requirements.txt
blinker==1.6.2
click==8.1.3
DateTime==5.1
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
python-dotenv==1.0.0
pytz==2023.3
Werkzeug==2.3.4
zope.interface==6.0
